chore: conceptually reordering the data structures of the generated commands within the domain on the project files
