+++ zork++/Cargo.toml
@@ -1,3 +1,4 @@
+workspace = { members = ["merging_toml"] }
 [package]
 name = "zork"
 version = "0.9.0"
