+++ zork++/src/lib/cli/output/mod.rs
@@ -1,3 +1,2 @@
 //! Defines operations or types that are related with send data to a system shell
-pub mod arguments;
-pub mod commands;
+pub mod executors;
