+++ zork++/src/lib/domain/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod arguments;
+pub mod command_lines;
