+++ zork++/src/lib/domain/mod.rs
@@ -1,2 +1,3 @@
+pub mod commands;
 pub mod target;
 pub mod translation_unit;
