+++ zork++/src/lib/domain/translation_unit.rs
@@ -1,20 +1,13 @@
 //! The module which holds the higher and generic abstractions over a source file
 
 use crate::project_model::compiler::StdLibMode;
+use color_eyre::Report;
+use serde::{Deserialize, Serialize};
 use std::borrow::Cow;
 use std::fmt::{Debug, Display};
 use std::path::PathBuf;
+use std::process::ExitStatus;
 use transient::{Any, Inv};
-
-/// The different type of translation units that `Zork++` is able to deal with
-#[derive(Debug)]
-pub enum TranslationUnitKind {
-    ModuleInterface,
-    ModuleImplementation,
-    SourceFile,
-    ModularStdLib(StdLibMode),
-    SystemHeader,
-}
 
 /// Represents any kind of translation unit and the generic operations
 /// applicable to all the implementors
@@ -95,4 +88,67 @@ macro_rules! impl_translation_unit_for {
             }
         }
     };
+}
+
+/// The different type of translation units that `Zork++` is able to deal with
+#[derive(Debug)]
+pub enum TranslationUnitKind {
+    ModuleInterface,
+    ModuleImplementation,
+    SourceFile,
+    ModularStdLib(StdLibMode),
+    SystemHeader,
+}
+
+/// The different states of a translation unit in the whole lifecycle of
+/// the build process and across different iterations of the same
+#[derive(Debug, Default, Serialize, Deserialize, Clone, Copy, PartialEq)]
+pub enum TranslationUnitStatus {
+    /// A command that is executed correctly
+    Success,
+    /// A skipped command due to previous successful iterations
+    Cached,
+    /// A command which is return code indicates an unsuccessful execution
+    Failed,
+    /// Whenever a translation unit must be rebuilt
+    #[default]
+    PendingToBuild,
+    /// The associated [`TranslationUnit`] has been deleted from the user's configuration and therefore,
+    /// it should be removed from the cache as well as its generated byproducts
+    ToDelete,
+    /// The execution failed, returning a [`color_eyre::Result`] with the Err variant
+    Error,
+}
+
+impl From<color_eyre::Result<ExitStatus, Report>> for TranslationUnitStatus {
+    fn from(value: color_eyre::Result<ExitStatus, Report>) -> Self {
+        helpers::handle_command_execution_result(&value)
+    }
+}
+
+impl From<&color_eyre::Result<ExitStatus, Report>> for TranslationUnitStatus {
+    fn from(value: &color_eyre::Result<ExitStatus, Report>) -> Self {
+        helpers::handle_command_execution_result(value)
+    }
+}
+
+mod helpers {
+    use crate::domain::translation_unit::TranslationUnitStatus;
+    use std::process::ExitStatus;
+
+    /// Convenient way of handle a command execution result avoiding duplicate code
+    pub(crate) fn handle_command_execution_result(
+        value: &color_eyre::Result<ExitStatus>,
+    ) -> TranslationUnitStatus {
+        match value {
+            Ok(r) => {
+                if r.success() {
+                    TranslationUnitStatus::Success
+                } else {
+                    TranslationUnitStatus::Failed
+                }
+            }
+            Err(_) => TranslationUnitStatus::Error,
+        }
+    }
 }
